backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sse-starlette
ffmpeg-python
opencv-python
scikit-image
boto3
requests
supabase
python-dotenv
ultralytics
sqlalchemy
pydantic-settings
psycopg2-binary
bcrypt
python-jose[cryptography]
